Divided creative.js into modules
